MLK-20216-1: drm/imx/dcss: fix suspicious sign extension
This addresses Coverity issues related to "Suspicious sign extension"
when an u16 is promoted to int (32bit signed) and then to u64. If the
resulting int is greater than 0x7fffffff the upper bits of the u64 will
all be 1.

+++ drivers/gpu/imx/dcss/dcss-dpr.c
@@ -390,7 +390,7 @@ void dcss_dpr_addr_set(struct dcss_soc *dcss, int ch_num, u32 luma_base_addr,
 		pitch = ch->pitch;
 
 	ch->frame_ctrl &= ~PITCH_MASK;
-	ch->frame_ctrl |= ((pitch << PITCH_POS) & PITCH_MASK);
+	ch->frame_ctrl |= (((u32)pitch << PITCH_POS) & PITCH_MASK);
 }
 EXPORT_SYMBOL(dcss_dpr_addr_set);
